Função de verificar o id para deixar de seguir

+++ cookenu1/src/index.ts
@@ -181,6 +181,12 @@ app.delete("/user/unfollow", async (req: Request, res: Response) => {
     const id_follower = authenticationData.id;
     const id_followed = req.body.userToUnfollowId
 
+    const followId = new FollowDatabase();
+    const idDb = await followId.isValidId(id_followed);
+    
+    if (idDb.quantity === 0) {
+      throw new Error("Insira um id válido");
+    }
     const followDb = new FollowDatabase();   
     await followDb.deleteFollow(id_followed, id_follower);
